+++ js/rpg_managers.js
@@ -2568,16 +2568,7 @@ BattleManager.makeActionOrders = function() {
 };
 
 BattleManager.startAction = function() {
-    var subject = this._subject;
-    var action = subject.currentAction();
-    var targets = action.makeTargets();
-    this._phase = 'action';
-    this._action = action;
-    this._targets = targets;
-    subject.useItem(action.item());
-    this._action.applyGlobal();
-    this.refreshStatus();
-    this._logWindow.startAction(subject, action, targets);
+    // YEP_BattleEngineCore.js
 };
 
 BattleManager.updateAction = function() {